Added ContentPermission Manager
