+++ protected/humhub/modules/user/tests/codeception/unit/GroupPermissionManagerMock.php
@@ -14,9 +14,9 @@ use humhub\modules\admin\permissions\ManageSettings;
 use humhub\modules\admin\permissions\ManageSpaces;
 use humhub\modules\admin\permissions\ManageUsers;
 use humhub\modules\admin\permissions\SeeAdminInformation;
-use humhub\modules\user\components\PermissionManager;
+use humhub\modules\user\components\GroupPermissionManager;
 
-class PermissionManagerMock extends PermissionManager
+class GroupPermissionManagerMock extends GroupPermissionManager
 {
     public $permissions = [
         2 => [
